Update flaky test documentation and add detailed view in advanced CI docs.
The flaky test documentation in the advanced CI section has been updated to include information about the "New flaky tests" and "Disruptions caused by a flaky test" charts. The "New flaky tests" chart now has the option to display data in a cumulative view by clicking the "Cumulative overview" button. Similarly, the "Disruptions caused by a flaky test" chart defaults to a cumulative overview but can be switched to a daily overview by clicking the "Daily overview" button.

In the detailed view, users now have access to the same set of actions as in the main view. They can add labels, mark a flaky test as resolved or unresolved, and include a link to the ticket related to the test. Additionally, aggregated data is provided, such as the P95 runtime, total number of runs, and pass rate of the test. Users can also navigate to specific jobs where the test diverged and analyze the wider view of the Disruption History chart. By default, the detailed view encompasses data from all branches.

Some known issues have been mentioned, such as the need to search before saving a new filter, otherwise, the filter contents may not be saved correctly. Another issue is that deleting a filter does not automatically switch to the next filter on the list.